+++ objects/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["animations", "entity", "player", "utils"]
+
+from . import *
